feat: Initial setup for Scraper Service

+++ src/main/resources/db/migration/V2__Add_Store_Relation.sql
@@ -1,2 +1,2 @@
-ALTER TABLE products ADD COLUMN store_id INTEGER;
+ALTER TABLE products ADD COLUMN store_id BIGINT;
 ALTER TABLE products ADD CONSTRAINT fk_product_store FOREIGN KEY  (store_id) REFERENCES stores (id);
